XFS design documentation
All technical documentation relating to the design and architecture
of the XFS filesystem will end up in the design/ directory. As a
first step, bring the delayed logging design documentation in the
kernel directory over and convert it to asciidoc format.

This gives us a build target to start adding makefile infrastructure
around to build the documentations. Initially, just start with
generating PDF format documentation.

Signed-off-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Dave Chinner <david@fromorbit.com>