docs: update refcount btree to reflect new cow staging extent policy
We now record the extents being used to stage a copy on write operation
in the refcount btree with the high bit of rc_startblock set and a
reference count of 1.  Setting the high bit helps us reduce CoW GC time
at mount by moving all the records to the right side of the tree.

Signed-off-by: Darrick J. Wong <darrick.wong@oracle.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>