docs: clarify what the ATTRBIT and ATTR2BIT feature flags do
It would seem that the disk format doesn't actually change depending on
ATTRBIT and ATTR2BIT.  The key difference is that ATTR2BIT (without the
'noattr2' mount option) allows di_forkoff to change depending on how
much space the attr fork needs, whereas ATTRBIT (without ATTR2BIT or
the 'attr2' mount option) does not allow di_forkoff to change.

In other words, at least one of the bits must be set if there are
extended attributes, but between the two bits the only difference is a
change in the default di_forkoff adjustment behaviors.

Signed-off-by: Darrick J. Wong <darrick.wong@oracle.com>